+++ src/Deli.Patcher/src/PatcherStage.cs
@@ -53,7 +53,7 @@ namespace Deli.Patcher
 			return null;
 		}
 
-		protected override Dictionary<string, AssetLoaderID>? GetAssets(Mod.AssetTable table)
+		protected override Mod.Asset[]? GetAssets(Mod.AssetTable table)
 		{
 			return table.Patcher;
 		}